[FEATURE] Added Fade Out / Fade In Animation for FOAB in ShoppingListDetail on Scroll. Added Recipe Functionality.

+++ lib/models/Item.dart
@@ -32,7 +32,7 @@ class Item {
       category: data["category"],
       count: data["count"],
       bought: data["bought"],
-      prize: data["prize"],
+      prize: data["prize"].toDouble(),
     );
   }
 
@@ -44,7 +44,7 @@ class Item {
       category: data["category"],
       count: data["count"],
       bought: data["bought"],
-      prize: data["prize"],
+      prize: data["prize"].toDouble(),
     );
   }
